+++ e2e/test-env/src/client/consumersThemes/themes.jsx
@@ -1,0 +1,3 @@
+import blogTheme from './blog/BlogTheme';
+
+export const themes = { blog: blogTheme };