+++ CHANGELOG.md
@@ -23,7 +23,10 @@
 ### :book: Documentation
 - `emoji`
   - [#1633](https://github.com/wix-incubator/rich-content/pull/1633) Emoji Plugin Docs
-  
+
+### 🏠 Internal
+- `e2e-tests`
+  - [#1681](https://github.com/wix-incubator/rich-content/pull/1681) Consumers themes(only Blog for now) added to RicosTestApp
 </details>
 <hr/>